requirements.txt
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
googletrans==4.0.2
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
sniffio==1.3.1
urllib3==2.3.0
